moved deprecate function to misc module
