+++ pyjob/misc.py
@@ -23,19 +23,13 @@
 __author__ = 'Felix Simkovic'
 __version__ = '1.0'
 
-import chardet.universaldetector
+import chardet
 import os
 import sys
 import tempfile
+import warnings
 
-from pyjob.deprecate import deprecate
 from pyjob.exception import PyJobError
-
-
-@deprecate(0.3, msg='use pyjob.stopwatch.StopWatch')
-def StopWatch():
-    from pyjob.stopwatch import StopWatch
-    return StopWatch()
 
 
 def decode(s):
@@ -63,15 +57,125 @@ def decode(s):
         if detector.done:
             break
     detector.close()
-    if detector.result['confidence'] < 0.98:
+    if not detector.done:
         raise PyJobError('Unable to infer string encoding')
     return s.decode(detector.result['encoding'])
+
+
+def deprecate(version, msg=None):
+    """Decorator to deprecate Python classes and functions
+    
+    Parameters
+    ----------
+    version : str
+       A string containing the version with which the callable is removed
+    msg : str, optional
+       An additional message that will be displayed alongside the default message
+
+    
+    Examples
+    --------
+    Enable :obj:`~DeprecationWarning` messages to be displayed.
+
+    >>> import warnings
+    >>> warnings.simplefilter('default')
+    
+    Decorate a simple Python function without additional message
+    
+    >>> @deprecate('0.0.0')
+    ... def sum(a, b):
+    ...     return a + b
+    >>> sum(1, 2)
+    deprecated.py:34: DeprecationWarning: sum has been deprecated and will be removed in version 0.0.0!
+      warnings.warn(message, DeprecationWarning)
+    3
+
+    Decorate a simple Python function with additional message
+
+    >>> @deprecate('0.0.1', msg='Use XXX instead!')
+    ... def sum(a, b):
+    ...     return a + b
+    >>> sum(2, 2)
+    deprecated.py:34: DeprecationWarning: sum has been deprecated and will be removed in version 0.0.0! - Use XXX instead!
+      warnings.warn(message, DeprecationWarning)
+    4
+
+    Decorate an entire Python class
+
+    >>> @deprecate('0.0.2')
+    ... class Obj(object):
+    ...     pass
+    >>> Obj()
+    deprecated.py:34: DeprecationWarning: Obj has been deprecated and will be removed in version 0.0.2!
+      warnings.warn(message, DeprecationWarning)
+    <__main__.Obj object at 0x7f8ee0f1ead0>
+
+    Decorate a Python class method
+
+    >>> class Obj(object):
+    ...     def __init__(self, v):
+    ...         self.v = v
+    ...     @deprecate('0.0.3')
+    ...     def mul(self, other):
+    ...         return self.v * other.v
+    >>> Obj(2).mul(Obj(3))
+    deprecated.py:34: DeprecationWarning: mul has been deprecated and will be removed in version 0.0.3!
+      warnings.warn(message, DeprecationWarning)
+    6
+
+    Decorate a Python class staticmethod
+    
+    >>> class Obj(object):
+    ...     @staticmethod
+    ...     @deprecate('0.0.4')
+    ...     def sub(a, b):
+    ...         return a - b
+    ... 
+    >>> Obj.sub(2, 1)
+    deprecated.py:34: DeprecationWarning: sub has been deprecated and will be removed in version 0.0.4!
+      warnings.warn(message, DeprecationWarning)
+    1
+
+    Decorate a Python class classmethod
+
+    >>> class Obj(object):
+    ...     CONST = 5
+    ...     @classmethod
+    ...     @deprecate('0.0.5')
+    ...     def sub(cls, a):
+    ...         return a - cls.CONST
+    ... 
+    >>> Obj().sub(5)
+    deprecated.py:34: DeprecationWarning: sub has been deprecated and will be removed in version 0.0.5!
+      warnings.warn(message, DeprecationWarning)
+    0
+
+    """
+
+    def deprecate_decorator(callable_):
+        def warn(*args, **kwargs):
+            message = "%s has been deprecated and will be removed in version %s!" % (
+                callable_.__name__, version)
+            if msg:
+                message += " - %s" % msg
+            warnings.warn(message, DeprecationWarning)
+            return callable_(*args, **kwargs)
+
+        return warn
+
+    return deprecate_decorator
 
 
 @deprecate(0.3, msg='use pyjob.stopwatch.StopWatch')
 def is_script(f):
     from pyjob.script import is_valid_script_path
     return is_valid_script_path(f)
+
+
+@deprecate(0.3, msg='use pyjob.stopwatch.StopWatch')
+def StopWatch():
+    from pyjob.stopwatch import StopWatch
+    return StopWatch()
 
 
 @deprecate(0.3, msg='use pyjob.script.Script')
